osmApi: refactoring + add way nodes
