+++ src/services/osm/auth/stringifyDomXml.ts
@@ -1,0 +1,8 @@
+import { isString } from '../../../components/helpers';
+
+export const stringifyDomXml = (itemXml: Node) => {
+  if (isString(itemXml)) {
+    throw new Error('String given');
+  }
+  return new XMLSerializer().serializeToString(itemXml);
+};
