+++ src/services/osm/osmToFeature.ts
@@ -20,6 +20,7 @@ export const osmToFeature = (element): Feature => {
     osmMeta,
     tags,
     members,
+    nodes, // TODO find a way to not pass it between SSR and client
     imageDefs: getImageDefs(tags, osmMeta.type, center),
     properties: { ...getPoiClass(tags) }, // TODO PoiClass is computed from tags, this can be removed
     deleted: osmappDeletedMarker,
